requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==11.0.0
torch==2.3.0
torchvision==0.18.0
